Conflicts:
	common/src/main/java/slash/common/io/Files.java
	datasource/src/main/java/slash/navigation/datasources/DataSourceManager.java
	datasource/src/main/java/slash/navigation/datasources/DataSourceService.java
	datasource/src/main/java/slash/navigation/datasources/DataSourcesUtil.java
	datasource/src/main/java/slash/navigation/datasources/File.java
	datasource/src/main/java/slash/navigation/datasources/impl/FileImpl.java
	datasource/src/main/java/slash/navigation/datasources/impl/MapImpl.java
	datasource/src/test/java/slash/navigation/datasources/DataSourceServiceTest.java
	download-tools/pom.xml
	download-tools/src/main/doc/create-datasources-xml.bat
	download/src/main/java/slash/navigation/download/Download.java
	download/src/main/java/slash/navigation/download/DownloadExecutor.java
	download/src/main/java/slash/navigation/download/DownloadExecutorComparator.java
	download/src/main/java/slash/navigation/download/DownloadManager.java
	download/src/main/java/slash/navigation/download/actions/Extractor.java
	download/src/main/java/slash/navigation/download/actions/Validator.java
	download/src/main/java/slash/navigation/download/datasources/File.java
	download/src/main/java/slash/navigation/download/datasources/Fragment.java
	download/src/main/java/slash/navigation/download/datasources/Map.java
	feedback/src/main/java/slash/navigation/feedback/domain/RouteFeedback.java
	feedback/src/test/java/slash/navigation/feedback/domain/ChecksumsIT.java
	feedback/src/test/java/slash/navigation/feedback/domain/RouteFeedbackServiceBase.java
	hgt/src/main/java/slash/navigation/hgt/HgtFiles.java
	hgt/src/test/java/slash/navigation/hgt/HgtFilesIT.java
	mapsforge-maps/src/main/java/slash/navigation/maps/models/OnlineMap.java
	mapsforge-maps/src/main/java/slash/navigation/maps/models/RemoteMapImpl.java
	mapsforge-maps/src/main/java/slash/navigation/maps/models/VectorMap.java
	rest/pom.xml
	rest/src/main/java/slash/navigation/rest/ReadRequest.java
	route-converter/src/main/java/slash/navigation/converter/gui/RouteConverter.java
	route-converter/src/main/java/slash/navigation/converter/gui/helpers/BatchPositionAugmenter.java
	route-converter/src/main/java/slash/navigation/converter/gui/helpers/RoutingServiceFacade.java
	route-converter/src/main/resources/slash/navigation/converter/gui/RouteConverter.properties

+++ download/src/main/java/slash/navigation/download/DownloadExecutor.java
@@ -36,6 +36,7 @@ import static java.lang.String.format;
 import static java.util.logging.Logger.getLogger;
 import static slash.common.io.Directories.ensureDirectory;
 import static slash.common.io.Files.setLastModified;
+import static slash.common.type.CompactCalendar.fromMillis;
 import static slash.navigation.download.State.*;
 
 /**
@@ -73,28 +74,41 @@ public class DownloadExecutor implements Runnable {
                 success = download();
             if (success) {
                 if (postProcess())
-                    updateState(download, Succeeded);
-            } else
-                updateState(download, Failed);
-            if (get.isNotModified())
+                    succeeded(get.isNotModified());
+
+            } else if (get.isNotModified())
                 updateState(download, NotModified);
+            else
+                failed();
         } catch (Exception e) {
             e.printStackTrace();
             log.severe(format("Could not download content from %s: %s", download.getUrl(), e));
-            updateState(download, Failed);
+            failed();
         }
     }
 
     private boolean canResume() {
-        Checksum latestChecksum = download.getFileChecksum();
+        Checksum checksum = download.getFileChecksum();
         return download.getTempFile().exists() && download.getTempFile().length() > 0 &&
-                latestChecksum != null && latestChecksum.getContentLength() != null &&
-                latestChecksum.getContentLength() > download.getTempFile().length();
+                checksum != null && checksum.getContentLength() != null &&
+                checksum.getContentLength() > download.getTempFile().length();
     }
 
     private void updateState(Download download, State state) {
         download.setState(state);
         downloadManager.getModel().updateDownload(download);
+        log.fine(format("State for download from %s changed to %s", download.getUrl(), state));
+    }
+
+    private void failed() {
+        updateState(download, Failed);
+        downloadManager.fireDownloadFailed(download);
+    }
+
+    private void succeeded(boolean notModified) {
+        updateState(download, Succeeded);
+        if (!notModified)
+            downloadManager.fireDownloadSucceeded(download);
     }
 
     private class ModelUpdater implements CopierListener {
@@ -105,6 +119,7 @@ public class DownloadExecutor implements Runnable {
         public void processedBytes(long byteCount) {
             download.setProcessedBytes(byteCount);
             downloadManager.getModel().updateDownload(download);
+            downloadManager.fireDownloadProgressed(download);
         }
     }
 
@@ -120,6 +135,7 @@ public class DownloadExecutor implements Runnable {
         get.setRange(fileSize, contentLength);
 
         InputStream inputStream = get.executeAsStream();
+        log.info(format("Resume from %s returned with status code %s", download.getUrl(), get.getStatusCode()));
         if (get.isPartialContent()) {
             modelUpdater.expectingBytes(contentLength != null ? contentLength : get.getContentLength() != null ? get.getContentLength() : 0);
             new Copier(modelUpdater).copyAndClose(inputStream, new FileOutputStream(download.getTempFile(), true), fileSize, contentLength);
@@ -138,6 +154,7 @@ public class DownloadExecutor implements Runnable {
             get.setIfNoneMatch(download.getETag());
 
         InputStream inputStream = get.executeAsStream();
+        log.info(format("Download from %s returned with status code %s", download.getUrl(), get.getStatusCode()));
         if (get.isSuccessful() && inputStream != null) {
             modelUpdater.expectingBytes(contentLength != null ? contentLength : get.getContentLength() != null ? get.getContentLength() : 0);
             new Copier(modelUpdater).copyAndClose(inputStream, new FileOutputStream(download.getTempFile()), 0, contentLength);
@@ -166,8 +183,11 @@ public class DownloadExecutor implements Runnable {
         if (!validate())
             return false;
 
-        if (!download.getTempFile().delete())
-            throw new IOException(format("Cannot delete temp file %s", download.getTempFile()));
+        if (download.getTempFile().exists())
+            if (!download.getTempFile().delete())
+                throw new IOException(format("Cannot delete temp file %s", download.getTempFile()));
+
+        log.fine(format("Postprocess from %s successful: %s", download.getUrl(), get.isSuccessful()));
         return true;
     }
 
@@ -189,21 +209,21 @@ public class DownloadExecutor implements Runnable {
         File target = download.getFileTarget();
         ensureDirectory(target.getParent());
         new Copier(modelUpdater).copyAndClose(download.getTempFile(), target);
-        setLastModified(target, get.getLastModified());
+        setLastModified(target, fromMillis(get.getLastModified()));
         return true;
     }
 
     private boolean flatten() throws IOException {
         File target = download.getFileTarget();
         new Extractor(modelUpdater).flatten(download.getTempFile(), target);
-        setLastModified(download.getTempFile(), get.getLastModified());
+        setLastModified(download.getTempFile(), fromMillis(get.getLastModified()));
         return true;
     }
 
     private boolean extract() throws IOException {
         File target = download.getFileTarget();
         new Extractor(modelUpdater).extract(download.getTempFile(), target);
-        setLastModified(download.getTempFile(), get.getLastModified());
+        setLastModified(download.getTempFile(), fromMillis(get.getLastModified()));
         return true;
     }
